Fix incorrect redirect
Fixed incorrect redirect to landing page when trying to order a burger when not authenticated. This error happens only if the user has already successfully ordered a burger before. The fix consists of resetting the purchased state to false once a burger order has completed. This allows the user to be correctly redirected to the checkout page when unauthenticated orders are attempted.